Remove -werror as a compiler flag
Treating warnings as errors leads to the build
stopping upon the first set of errors. This means
multiple builds to discover all the warnings,
and a lack of forward progress when we have yet
to write the code that uses a variable, for example.
Warnings still appear in the compiler output, so
we are aware of them.