Use request_buffer_size in echoRequestHandler
By having a separate buffer size for the request_buffer,
we avoid reusing the response_buffer_size variable.
That avoids a contradictory usage of a variable
(that otherwise doesn't live up to its name),
and lets us make the request_buffer and its size
constant, to reflect their read-only nature.