Pass response_buffer pointer by reference
In order for a callee to use a pointer as an out-param,
it has to be able to modify the pointer itself (the
memory address stored, rather than what it points to).
We can accomplish this by using a pointer-to-a-pointer,
or more elegantly (as done here), by passing that
pointer by reference.