Allow routing to be set from the config file
Instead of passing in a config parameter name (for ex: "port"), we
pass in a path to the desired parameter. For the case of port, we
would use ["server", "listen"]. This allows us to use nested blocks
in our config file, which we use for setting up the routing in our
web server. This routing specifies the base url from which files
will be served as well as the Handler which will handle the
response.