Refactor ServerConfig using prefix to handler map
API Changes:

Changed the ServerConfig class to construct both maps of path to
config property value and prefix to handler names. A ServerConfig
object is initialized using Build() so that we can catch any errors
that might occur from reading a malformed config file.

Also introduced are two new getters, `allPaths()` and `getChildBlock`
which return the mapping of prefixes to handler name and inner
NginxConfig block of the specified route config respectively.

Other Changes:

Refactored `fillOutMaps()` so that we no longer make use of exceptions
when checking if a property exists in a map. Instead we use
`unordered_map::find()`.