Fix possible URI prefix generation
We are starting our substring search from the
end of the string (so from size() - 1 to 0), but
we want to cut from the start to the found forward-slash.
For example:

h e l / l o /
0 1 2 3 4 5 6
  ^   ^
  |   |
  |   pos
  |
  search_index

We start our search from search_index (index 1), but find the
forward-slash at pos (index 3). The prefix we want to consider
is "hel", the substring starting at 0 and with length 3 (including
the first character, so the trailing slash is not included).
Thus, the call to make is: substr(0, pos), where pos is found
by a search starting from search_index.